fix: ahci - no uint32 truncation on volumes > 2TiB (bug-hunt #287)
Parse the 48-bit LBA capacity from IDENTIFY words 100-103 instead of
only words 100-101, so drives >= 2TiB report their real sector count
instead of silently wrapping (28-bit words 57-58 kept as fallback).
Widen struct ahci_port.sector_count, ahci_get_sectors(), and the
libc/userspace shim chain to uint64_t. Program the full 48-bit LBA in
the non-NCQ READ/WRITE DMA EXT path (was truncated to 32 bits), and
widen ahci_ncq_read/write lba params to uint64_t.

+++ src/include/ahci.h
@@ -53,7 +53,7 @@
 int  ahci_init(void);
 void ahci_exit(void);
 int  ahci_is_present(void);
-uint32_t ahci_get_sectors(void);
+uint64_t ahci_get_sectors(void);
 int  ahci_read_sectors(uint32_t lba, uint8_t count, void *buf);
 int  ahci_write_sectors(uint32_t lba, uint8_t count, const void *buf);
 
@@ -69,13 +69,11 @@ int  ahci_has_ncq_cap(void);
  * @count:    sector count (1-8 typically)
  * @buf:      data buffer (size = count * 512)
  * Returns 0 on success, -1 on error. */
-int  ahci_ncq_read(int port_num, int pm_port, uint32_t lba,
-                    uint8_t count, void *buf);
+int ahci_ncq_read(int port_num, int pm_port, uint64_t lba, uint8_t count, void *buf);
 
 /* Submit a queued NCQ WRITE FPDMA QUEUED command.
  * Same parameters as ahci_ncq_read. */
-int  ahci_ncq_write(int port_num, int pm_port, uint32_t lba,
-                     uint8_t count, const void *buf);
+int ahci_ncq_write(int port_num, int pm_port, uint64_t lba, uint8_t count, const void *buf);
 
 /* Poll for NCQ completions on a physical port.
  * Checks the SDB FIS (Set Device Bits) and completes finished commands.
