+++ userspace/kmods/shell/cmds/cmd_iostat.c
@@ -6,7 +6,7 @@
 
 /* Track previous sector counts to compute deltas */
 static uint32_t prev_ata_sectors = 0;
-static uint32_t prev_ahci_sectors = 0;
+static uint64_t prev_ahci_sectors = 0;
 
 void cmd_iostat(const char *args) {
     (void)args;
@@ -14,7 +14,7 @@ void cmd_iostat(const char *args) {
     kprintf("Device              Reads(sects)   Delta\n");
 
     uint32_t ata_sects = libc_ata_is_present() ? libc_ata_get_sectors() : 0;
-    uint32_t ahci_sects = libc_ahci_is_present() ? libc_ahci_get_sectors() : 0;
+    uint64_t ahci_sects = libc_ahci_is_present() ? libc_ahci_get_sectors() : 0;
 
     if (libc_ata_is_present()) {
         int64_t delta = (int64_t)(ata_sects - prev_ata_sectors);
@@ -23,7 +23,8 @@ void cmd_iostat(const char *args) {
     }
     if (libc_ahci_is_present()) {
         int64_t delta = (int64_t)(ahci_sects - prev_ahci_sectors);
-        kprintf("ahci0               %-15u %lld\n", ahci_sects, (long long)delta);
+        kprintf("ahci0               %-15llu %lld\n", (unsigned long long)ahci_sects,
+                (long long)delta);
         prev_ahci_sectors = ahci_sects;
     }
     kprintf("\nNote: readings show total sector count (monotonically increasing)\n");
